v0.9.7b
Added:
Added New logo to demo page.
Added `navbar-center` to center a logo in the navbar, this was originally automatically centered but created issues with gaps on the sides.

Fixed:
Fixed height issues with the navbar, this was caused by vertical alignment of a flex item which removed height from automatically being assigned to 100%.
Fixed navbar dropdown format on the top of the demo.